requirements.txt
Django==2.0.4
django-extensions==2.0.7
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.7.1
pytz==2018.4
six==1.11.0
